+++ modules/core/src/index.ts
@@ -84,7 +84,13 @@ export type {PipelineLayoutProps} from './adapter/resources/pipeline-layout';
 export {PipelineLayout} from './adapter/resources/pipeline-layout';
 
 // PORTABLE API - UNIFORM BUFFERS
-export {UniformBufferLayout} from './portable/uniform-buffer-layout';
+export {
+  makeShaderBlockLayout,
+  type ShaderBlockLayout,
+  type ShaderBlockLayoutEntry,
+  type ShaderBlockLayoutOptions
+} from './shadertypes/shader-types/shader-block-layout';
+export {ShaderBlockWriter} from './portable/shader-block-writer';
 export {UniformBlock} from './portable/uniform-block';
 export {UniformStore} from './portable/uniform-store';
 // TEXTURE TYPES